+++ crates/interpreter/src/lib.rs
@@ -226,7 +226,7 @@ impl InterpreterState {
 
     pub fn trigger_exception(&mut self, exception: Exception, exception_handler: Address) {
         self.registers.vaddr = exception.vaddr().unwrap_or_default();
-        self.registers.status |= 0x00000002; // sets bit 1
+        self.registers.status.set_exception_level(true);
         self.registers.cause &= 0xffffff83; // clears bits 2-6
         self.registers.cause |= exception.code() << 2;
         self.registers.epc = self.pc - 4;